:hammer: add a database object
