+++ src/api/ddb/index.js
@@ -1,0 +1,3 @@
+const ddb = require('./ddb')
+
+module.exports = ddb
