+++ src/api/index.js
@@ -1,6 +1,6 @@
 const bodyParser = require('body-parser')
 const { API_BASE_CONTEXT } = require('../constants')
-const state = require('./state')
+const state = require('./models/state')
 
 const genPath = resource => `${API_BASE_CONTEXT}/${resource.NAME}`
