+++ src/api/models/state/constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+  NAME: 'state',
+}
